Add data-cy to navcard and updated test
