+++ src/StockportWebapp/Views/stockportgov/Home/Index2023.cshtml
@@ -48,7 +48,7 @@
 
 <section class="section section--white">
     <div class="grid-container">
-        <div class="grid-100" data-test="navcard-list">
+        <div class="grid-100" data-cy="navcard-list">
             <h2 class="section__title section__title--padded-bottom h-l">Our services</h2>
             <partial name="NavCardList" model='Model.HomepageContent.Services'/>
