+++ src/StockportWebapp/Views/stockportgov/Shared/NavCard.cshtml
@@ -1,5 +1,5 @@
 ﻿@model NavCard
-<a href="@Model.Url">
+<a href="@Model.Url" data-cy="nav-card-link">
     <h3 class="navcard__title">@Model.Title</h3>
     <p class="navcard__teaser">@Model.Teaser</p>
 </a>
